Move product queries and mutations to a separate file
