+++ saleor/graphql/api.py
@@ -1,8 +1,7 @@
 import graphene
 import graphql_jwt
 from graphene_django.filter import DjangoFilterConnectionField
-from graphql_jwt.decorators import (
-    login_required, permission_required, staff_member_required)
+from graphql_jwt.decorators import login_required, permission_required
 
 from .account.mutations import (
     CustomerCreate, CustomerUpdate, CustomerPasswordReset, CustomerRegister,
@@ -39,63 +38,22 @@ from .order.mutations.orders import (
 from .page.resolvers import resolve_pages, resolve_page
 from .page.types import Page
 from .page.mutations import PageCreate, PageDelete, PageUpdate
-from .product.filters import ProductFilterSet
-from .product.mutations.attributes import (
-    AttributeValueCreate, AttributeValueDelete,
-    AttributeValueUpdate, AttributeCreate, AttributeDelete,
-    AttributeUpdate)
-from .product.mutations.products import (
-    CategoryCreate, CategoryDelete, CategoryUpdate,
-    CollectionAddProducts, CollectionCreate, CollectionDelete,
-    CollectionRemoveProducts, CollectionUpdate, ProductCreate,
-    ProductDelete, ProductUpdate, ProductTypeCreate,
-    ProductTypeDelete, ProductImageCreate, ProductImageDelete,
-    ProductImageReorder, ProductImageUpdate, ProductTypeUpdate,
-    ProductVariantCreate, ProductVariantDelete,
-    ProductVariantUpdate, VariantImageAssign, VariantImageUnassign)
-from .product.resolvers import (
-    resolve_attributes, resolve_categories, resolve_collections,
-    resolve_products, resolve_product_types, resolve_product_variants,
-    resolve_report_product_sales)
-from .product.types import (
-    ReportingPeriod, Category, Collection, Product, Attribute, ProductType,
-    ProductVariant, ProductSales)
+from .product.schema import ProductMutations, ProductQueries
 from .shipping.resolvers import resolve_shipping_zones
 from .shipping.types import ShippingZone
 from .shipping.mutations import (
     ShippingZoneCreate, ShippingZoneDelete, ShippingZoneUpdate,
     ShippingPriceCreate, ShippingPriceDelete, ShippingPriceUpdate)
-
 from .shop.types import Shop
 from .shop.mutations import (
     AuthorizationKeyAdd, AuthorizationKeyDelete, ShopDomainUpdate,
     ShopSettingsUpdate, HomepageCollectionUpdate)
 
 
-class Query(graphene.ObjectType):
+class Query(ProductQueries):
     address_validator = graphene.Field(
         AddressValidationData,
         input=graphene.Argument(AddressValidationInput, required=True))
-    attributes = DjangoFilterConnectionField(
-        Attribute,
-        query=graphene.String(description=DESCRIPTIONS['attributes']),
-        in_category=graphene.Argument(graphene.ID),
-        description='List of the shop\'s attributes.')
-    categories = DjangoFilterConnectionField(
-        Category, query=graphene.String(
-            description=DESCRIPTIONS['category']),
-        level=graphene.Argument(graphene.Int),
-        description='List of the shop\'s categories.')
-    category = graphene.Field(
-        Category, id=graphene.Argument(graphene.ID),
-        description='Lookup a category by ID.')
-    collection = graphene.Field(
-        Collection, id=graphene.Argument(graphene.ID),
-        description='Lookup a collection by ID.')
-    collections = DjangoFilterConnectionField(
-        Collection, query=graphene.String(
-            description=DESCRIPTIONS['collection']),
-        description='List of the shop\'s collections.')
     menu = graphene.Field(
         Menu, id=graphene.Argument(graphene.ID),
         name=graphene.Argument(graphene.String, description="Menu name."),
@@ -123,24 +81,6 @@ class Query(graphene.ObjectType):
         Page, query=graphene.String(
             description=DESCRIPTIONS['page']),
         description='List of the shop\'s pages.')
-    product = graphene.Field(
-        Product, id=graphene.Argument(graphene.ID),
-        description='Lookup a product by ID.')
-    products = DjangoFilterConnectionField(
-        Product, filterset_class=ProductFilterSet, query=graphene.String(
-            description=DESCRIPTIONS['product']),
-        description='List of the shop\'s products.')
-    product_type = graphene.Field(
-        ProductType, id=graphene.Argument(graphene.ID),
-        description='Lookup a product type by ID.')
-    product_types = DjangoFilterConnectionField(
-        ProductType, description='List of the shop\'s product types.')
-    product_variant = graphene.Field(
-        ProductVariant, id=graphene.Argument(graphene.ID),
-        description='Lookup a variant by ID.')
-    product_variants = DjangoFilterConnectionField(
-        ProductVariant, ids=graphene.List(graphene.ID),
-        description='Lookup multiple variants by ID')
     sale = graphene.Field(
         Sale, id=graphene.Argument(graphene.ID),
         description='Lookup a sale by ID.')
@@ -148,10 +88,6 @@ class Query(graphene.ObjectType):
         Sale, query=graphene.String(description=DESCRIPTIONS['sale']),
         description="List of the shop\'s sales.")
     shop = graphene.Field(Shop, description='Represents a shop resources.')
-    report_product_sales = DjangoFilterConnectionField(
-        ProductSales,
-        period=graphene.Argument(ReportingPeriod, required=True),
-        description='List of top selling products.')
     voucher = graphene.Field(
         Voucher, id=graphene.Argument(graphene.ID),
         description='Lookup a voucher by ID.')
@@ -174,21 +110,6 @@ class Query(graphene.ObjectType):
         User, description='List of the shop\'s staff users.',
         query=graphene.String(description=DESCRIPTIONS['user']))
     node = graphene.Node.Field()
-
-    def resolve_attributes(self, info, in_category=None, query=None, **kwargs):
-        return resolve_attributes(info, in_category, query)
-
-    def resolve_category(self, info, id):
-        return graphene.Node.get_node_from_global_id(info, id, Category)
-
-    def resolve_categories(self, info, level=None, query=None, **kwargs):
-        return resolve_categories(info, level=level, query=query)
-
-    def resolve_collection(self, info, id):
-        return graphene.Node.get_node_from_global_id(info, id, Collection)
-
-    def resolve_collections(self, info, query=None, **kwargs):
-        return resolve_collections(info, query)
 
     @permission_required(['account.manage_users'])
     def resolve_user(self, info, id):
@@ -228,18 +149,6 @@ class Query(graphene.ObjectType):
     def resolve_orders(self, info, query=None, **kwargs):
         return resolve_orders(info, query)
 
-    def resolve_product(self, info, id):
-        return graphene.Node.get_node_from_global_id(info, id, Product)
-
-    def resolve_products(self, info, category_id=None, query=None, **kwargs):
-        return resolve_products(info, category_id, query)
-
-    def resolve_product_type(self, info, id):
-        return graphene.Node.get_node_from_global_id(info, id, ProductType)
-
-    def resolve_product_types(self, info, **kwargs):
-        return resolve_product_types()
-
     def resolve_shop(self, info):
         return Shop()
 
@@ -250,16 +159,6 @@ class Query(graphene.ObjectType):
     @permission_required('discount.manage_discounts')
     def resolve_sales(self, info, query=None, **kwargs):
         return resolve_sales(info, query)
-
-    def resolve_product_variant(self, info, id):
-        return graphene.Node.get_node_from_global_id(info, id, ProductVariant)
-
-    def resolve_product_variants(self, info, ids=None, **kwargs):
-        return resolve_product_variants(info, ids)
-
-    @staff_member_required
-    def resolve_report_product_sales(self, info, period, **kwargs):
-        return resolve_report_product_sales(info, period)
 
     @permission_required('discount.manage_discounts')
     def resolve_voucher(self, info, id):
@@ -281,7 +180,7 @@ class Query(graphene.ObjectType):
         return resolve_address_validator(info, input)
 
 
-class Mutations(graphene.ObjectType):
+class Mutations(ProductMutations):
     authorization_key_add = AuthorizationKeyAdd.Field()
     authorization_key_delete = AuthorizationKeyDelete.Field()
 
@@ -293,14 +192,6 @@ class Mutations(graphene.ObjectType):
 
     set_password = SetPassword.Field()
     password_reset = PasswordReset.Field()
-
-    attribute_value_create = AttributeValueCreate.Field()
-    attribute_value_delete = AttributeValueDelete.Field()
-    attribute_value_update = AttributeValueUpdate.Field()
-
-    category_create = CategoryCreate.Field()
-    category_delete = CategoryDelete.Field()
-    category_update = CategoryUpdate.Field()
 
     customer_create = CustomerCreate.Field()
     customer_update = CustomerUpdate.Field()
@@ -314,12 +205,6 @@ class Mutations(graphene.ObjectType):
     address_create = AddressCreate.Field()
     address_update = AddressUpdate.Field()
     address_delete = AddressDelete.Field()
-
-    collection_create = CollectionCreate.Field()
-    collection_update = CollectionUpdate.Field()
-    collection_delete = CollectionDelete.Field()
-    collection_add_products = CollectionAddProducts.Field()
-    collection_remove_products = CollectionRemoveProducts.Field()
 
     menu_create = MenuCreate.Field()
     menu_delete = MenuDelete.Field()
@@ -352,27 +237,6 @@ class Mutations(graphene.ObjectType):
     page_delete = PageDelete.Field()
     page_update = PageUpdate.Field()
 
-    attribute_create = AttributeCreate.Field()
-    attribute_delete = AttributeDelete.Field()
-    attribute_update = AttributeUpdate.Field()
-
-    product_create = ProductCreate.Field()
-    product_delete = ProductDelete.Field()
-    product_update = ProductUpdate.Field()
-
-    product_image_create = ProductImageCreate.Field()
-    product_image_reorder = ProductImageReorder.Field()
-    product_image_delete = ProductImageDelete.Field()
-    product_image_update = ProductImageUpdate.Field()
-
-    product_type_create = ProductTypeCreate.Field()
-    product_type_update = ProductTypeUpdate.Field()
-    product_type_delete = ProductTypeDelete.Field()
-
-    product_variant_create = ProductVariantCreate.Field()
-    product_variant_delete = ProductVariantDelete.Field()
-    product_variant_update = ProductVariantUpdate.Field()
-
     sale_create = SaleCreate.Field()
     sale_delete = SaleDelete.Field()
     sale_update = SaleUpdate.Field()
@@ -392,9 +256,6 @@ class Mutations(graphene.ObjectType):
     shipping_price_create = ShippingPriceCreate.Field()
     shipping_price_delete = ShippingPriceDelete.Field()
     shipping_price_update = ShippingPriceUpdate.Field()
-
-    variant_image_assign = VariantImageAssign.Field()
-    variant_image_unassign = VariantImageUnassign.Field()
 
 
 schema = graphene.Schema(Query, Mutations)
